Uses main thread to perform cryptography stuff now
