+++ android/src/main/java/com/rnbiometrics/ReactNativeBiometricsPackage.java
@@ -15,10 +15,10 @@ import java.util.List;
  */
 
 public class ReactNativeBiometricsPackage implements ReactPackage {
-    @Override
-    public List<Class<? extends JavaScriptModule>> createJSModules() {
-        return Collections.emptyList();
-    }
+    // @Override
+    // public List<Class<? extends JavaScriptModule>> createJSModules() {
+    //     return Collections.emptyList();
+    // }
 
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
